+++ DietLog-MVVM/View/Meal/Controller/MealEditViewController.swift
@@ -12,23 +12,8 @@ import RxSwift
 class MealEditViewController: BaseViewController {
 
     // MARK: - Component
-    private lazy var mealEditView = MealEditView()
-    
-    // MARK: - 변수
-    private var viewModel = MealEditViewModel()
-    private var selectedDate = Date()
-    private var disposeBag = DisposeBag()
-    
-    // MARK: - 초기화
-    init(selectedDate: Date) {
-        super.init(nibName: nil, bundle: nil)
-        self.selectedDate = selectedDate
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
+    lazy var mealEditView = MealEditView()
+
     // MARK: - Life Cycle
     override func loadView() {
         super.loadView()
@@ -42,38 +27,10 @@ class MealEditViewController: BaseViewController {
         mealEditView.configure()
         mealEditView.delegate = self
     }
-    
-    // MARK: - Setup NavigationBar
-    override func setupNavigationBar() {
-        let button = UIBarButtonItem(title: "저장", style: .plain, target: self, action: #selector(saveMeal))
-        navigationItem.rightBarButtonItem = button
-        
-        navigationItem.title = "식단 쓰기"
-    }
-    
-    // MARK: - Setup Bind
-    override func setupBinding() {
-        mealEditView.memoTextView.rx.attributedText
-            .bind(to: viewModel.memoTextView)
-            .disposed(by: disposeBag)
-    }
 }
 
 // MARK: - 메서드
 extension MealEditViewController {
-
-    @objc func saveMeal() {
-        let result = viewModel.saveMeal(for: selectedDate, and: mealEditView.selectedImage)
-        
-        if !result {
-            showAlertWithOKButton(title: "", message: "글 입력 혹은 이미지 추가를 해주세요")
-        } else {
-            showAlertWithOKButton(title: "", message: "저장했습니다") {
-                self.navigationController?.popViewController(animated: true)
-            }
-        }
-    }
-
     private func insertImageIntoTextView(_ image: UIImage) {
         let memoTextViewWidth = mealEditView.memoTextView.frame.size.width
         let imageWidth = image.size.width
@@ -104,7 +61,6 @@ extension MealEditViewController {
     }
     
     private func isContainsImage(in textView: UITextView) -> Bool {
-        
         guard let attributedText = textView.attributedText else { return false }
         
         let range = NSRange(location: 0, length: attributedText.length)
